feat(memory): neurodivergent cognitive profiles, lateral retrieval, ICNU fusion
- Add CognitiveProfile with ADHD/ASD/Dyslexia support
- Implement neurodivergent package: HyperfocusState, LateralEvaluator, IcnuWeights
- Add SemanticRecallStrategy for context-aware retrieval
- Update pipelines, scorers, and persistence for config-driven architecture
- Comprehensive test coverage for all new subsystems

+++ spector-memory/src/main/java/com/spectrayan/spector/memory/hippocampus/TombstoneCompactor.java
@@ -81,7 +81,7 @@ public final class TombstoneCompactor {
             if (SynapticHeaderConstants.isPinned(flags)) continue;
 
             long timestamp = layout.readTimestamp(segment, offset);
-            short recallCount = layout.readRecallCount(segment, offset);
+            int recallCount = layout.readRecallCount(segment, offset);
             float importance = layout.readImportance(segment, offset);
 
             float decay = DecayStrategy.computeDecay(timestamp, nowMs, recallCount);
